agents/official/beeai-framework/rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "TBD"
authors = [
    { name = "IBM Corp." },
]
requires-python = ">=3.13,<4"
dependencies = [
    "beeai-framework[duckduckgo,wikipedia]~=0.1.36",
    "beeai-sdk",
    "openinference-instrumentation-beeai>=0.1.8",
    "pydantic-settings>=2.9.0",
    "tenacity>=9.1.2",
    "langchain-text-splitters>=0.3.0",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 120

[tool.uv.sources]
beeai-sdk = { path = "../../../../apps/beeai-sdk", editable = true }

[project.scripts]
server = "rag.agent:serve"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
